Fix iterator problems in TupleList, uncovered trying to sort. (2018-01-16)
eigen.hpp
=========

TupleList no longer subclasses from std::vector<>.  This was a
cardinal sin for C++ (and the iterators were failing on std::sort()).
The immediate consequence of this is to add methods for size(),
clear(), reserve(), operator[].

https://stackoverflow.com/questions/4353203/thou-shalt-not-inherit-from-stdvector

Even after the subclassing change, the iterator TupleList<>::iterator
works for Spsparse accumulators, but not for std::sort.  For sorting,
the user must dig deeper for TupleList.tuples.begin(), etc.  This is
inconvenient and messy, but it works.

TODO: Figure out how to fix this properly; i.e. iterators for
      TupleList that work for all purposes.

test_array.cpp
==============

Test added to sort a tuple list.